requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
djangorestframework==3.12.4
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
numpy==1.22.3
psycopg2==2.9.1
pytz==2021.3
redis==4.1.4
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
zope.event==4.5.0
zope.interface==5.4.0
